+++ src/control/cmd/daos_agent/mgmt_rpc.go
@@ -10,7 +10,6 @@ import (
 	"net"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/pkg/errors"
 	"golang.org/x/net/context"
@@ -44,12 +43,8 @@ type mgmtModule struct {
 	monitor        *procMon
 	useDefaultNUMA bool
 
-	numaGetter     hardware.ProcessNUMAProvider
-	providerIdx    uint
-	devClassGetter hardware.NetDevClassProvider
-	devStateGetter hardware.NetDevStateProvider
-	fabricScanner  *hardware.FabricScanner
-	netIfaces      func() ([]net.Interface, error)
+	numaGetter  hardware.ProcessNUMAProvider
+	providerIdx uint
 }
 
 func (mod *mgmtModule) HandleCall(ctx context.Context, session *drpc.Session, method drpc.Method, req []byte) ([]byte, error) {
@@ -338,42 +333,7 @@ func (mod *mgmtModule) getProviderIdxURIs(srvResp *mgmtpb.GetAttachInfoResp, idx
 }
 
 func (mod *mgmtModule) getFabricInterface(ctx context.Context, params *FabricIfaceParams) (*FabricInterface, error) {
-	if err := mod.waitFabricReady(ctx, params.DevClass); err != nil {
-		return nil, err
-	}
-
 	return mod.cache.GetFabricDevice(ctx, params)
-}
-
-func (mod *mgmtModule) waitFabricReady(ctx context.Context, netDevClass hardware.NetDevClass) error {
-	mod.fabricMutex.Lock()
-	defer mod.fabricMutex.Unlock()
-
-	if mod.netIfaces == nil {
-		mod.netIfaces = net.Interfaces
-	}
-	ifaces, err := mod.netIfaces()
-	if err != nil {
-		return err
-	}
-
-	var needIfaces []string
-	for _, iface := range ifaces {
-		devClass, err := mod.devClassGetter.GetNetDevClass(iface.Name)
-		if err != nil {
-			return err
-		}
-		if devClass == netDevClass {
-			needIfaces = append(needIfaces, iface.Name)
-		}
-	}
-
-	return hardware.WaitFabricReady(ctx, mod.log, hardware.WaitFabricReadyParams{
-		StateProvider:  mod.devStateGetter,
-		FabricIfaces:   needIfaces,
-		IgnoreUnusable: true,
-		IterationSleep: time.Second,
-	})
 }
 
 func (mod *mgmtModule) handleNotifyPoolConnect(ctx context.Context, reqb []byte, pid int32) error {
