+++ src/control/cmd/daos_agent/mgmt_rpc_test.go
@@ -202,23 +202,9 @@ func TestAgent_mgmtModule_getAttachInfo(t *testing.T) {
 				cache: newTestInfoCache(t, log, testInfoCacheParams{
 					mockGetAttachInfo: tc.mockGetAttachInfo,
 					mockScanFabric:    tc.mockFabricScan,
+					mockNetIfaces:     tc.mockGetNetIfaces,
 				}),
 				numaGetter: tc.numaGetter,
-				devClassGetter: &hardware.MockNetDevClassProvider{
-					GetNetDevClassReturn: []hardware.MockGetNetDevClassResult{
-						{
-							NDC: hardware.Ether,
-						},
-					},
-				},
-				devStateGetter: &hardware.MockNetDevStateProvider{
-					GetStateReturn: []hardware.MockNetDevStateResult{
-						{
-							State: hardware.NetDevStateReady,
-						},
-					},
-				},
-				netIfaces: tc.mockGetNetIfaces,
 			}
 
 			respBytes, err := mod.handleGetAttachInfo(test.Context(t), tc.reqBytes, 123)
@@ -290,20 +276,6 @@ func TestAgent_mgmtModule_getAttachInfo_Parallel(t *testing.T) {
 		sys:        sysName,
 		cache:      ic,
 		ctlInvoker: ctlInvoker,
-		devClassGetter: &hardware.MockNetDevClassProvider{
-			GetNetDevClassReturn: []hardware.MockGetNetDevClassResult{
-				{
-					NDC: hardware.Ether,
-				},
-			},
-		},
-		devStateGetter: &hardware.MockNetDevStateProvider{
-			GetStateReturn: []hardware.MockNetDevStateResult{
-				{
-					State: hardware.NetDevStateReady,
-				},
-			},
-		},
 	}
 
 	var wg sync.WaitGroup
@@ -379,117 +351,6 @@ func TestAgent_mgmtModule_getNUMANode(t *testing.T) {
 
 			test.AssertEqual(t, tc.expResult, result, "")
 			test.CmpErr(t, tc.expErr, err)
-		})
-	}
-}
-
-func TestAgent_mgmtModule_waitFabricReady(t *testing.T) {
-	defaultNetIfaceFn := func() ([]net.Interface, error) {
-		return []net.Interface{
-			{Name: "t0"},
-			{Name: "t1"},
-			{Name: "t2"},
-		}, nil
-	}
-
-	defaultDevClassProv := &hardware.MockNetDevClassProvider{
-		GetNetDevClassReturn: []hardware.MockGetNetDevClassResult{
-			{
-				ExpInput: "t0",
-				NDC:      hardware.Infiniband,
-			},
-			{
-				ExpInput: "t1",
-				NDC:      hardware.Infiniband,
-			},
-			{
-				ExpInput: "t2",
-				NDC:      hardware.Ether,
-			},
-		},
-	}
-
-	for name, tc := range map[string]struct {
-		netIfacesFn  func() ([]net.Interface, error)
-		devClassProv *hardware.MockNetDevClassProvider
-		devStateProv *hardware.MockNetDevStateProvider
-		netDevClass  hardware.NetDevClass
-		expErr       error
-		expChecked   []string
-	}{
-		"netIfaces fails": {
-			netIfacesFn: func() ([]net.Interface, error) {
-				return nil, errors.New("mock netIfaces")
-			},
-			netDevClass: hardware.Infiniband,
-			expErr:      errors.New("mock netIfaces"),
-		},
-		"GetNetDevClass fails": {
-			devClassProv: &hardware.MockNetDevClassProvider{
-				GetNetDevClassReturn: []hardware.MockGetNetDevClassResult{
-					{
-						ExpInput: "t0",
-						Err:      errors.New("mock GetNetDevClass"),
-					},
-				},
-			},
-			netDevClass: hardware.Infiniband,
-			expErr:      errors.New("mock GetNetDevClass"),
-		},
-		"GetNetDevState fails": {
-			devStateProv: &hardware.MockNetDevStateProvider{
-				GetStateReturn: []hardware.MockNetDevStateResult{
-					{Err: errors.New("mock NetDevStateProvider")},
-				},
-			},
-			netDevClass: hardware.Infiniband,
-			expErr:      errors.New("mock NetDevStateProvider"),
-			expChecked:  []string{"t0"},
-		},
-		"down devices are ignored": {
-			devStateProv: &hardware.MockNetDevStateProvider{
-				GetStateReturn: []hardware.MockNetDevStateResult{
-					{State: hardware.NetDevStateDown},
-					{State: hardware.NetDevStateReady},
-				},
-			},
-			netDevClass: hardware.Infiniband,
-			expChecked:  []string{"t0", "t1"},
-		},
-		"success": {
-			netDevClass: hardware.Infiniband,
-			expChecked:  []string{"t0", "t1"},
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			log, buf := logging.NewTestLogger(t.Name())
-			defer test.ShowBufferOnFailure(t, buf)
-
-			if tc.netIfacesFn == nil {
-				tc.netIfacesFn = defaultNetIfaceFn
-			}
-
-			if tc.devClassProv == nil {
-				tc.devClassProv = defaultDevClassProv
-			}
-
-			if tc.devStateProv == nil {
-				tc.devStateProv = &hardware.MockNetDevStateProvider{}
-			}
-
-			mod := &mgmtModule{
-				log:            log,
-				netIfaces:      tc.netIfacesFn,
-				devClassGetter: tc.devClassProv,
-				devStateGetter: tc.devStateProv,
-			}
-
-			err := mod.waitFabricReady(test.Context(t), tc.netDevClass)
-
-			test.CmpErr(t, tc.expErr, err)
-			if diff := cmp.Diff(tc.expChecked, tc.devStateProv.GetStateCalled); diff != "" {
-				t.Fatalf("-want, +got:\n%s", diff)
-			}
 		})
 	}
 }
